give permission to specific user

+++ app/Http/Controllers/Backend/RolesController.php
@@ -6,11 +6,21 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Contracts\Session\Session;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
 class RolesController extends Controller
 {
+    public $user;
+
+    public function __construct(){
+        $this->middleware(function($request, $next){
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -18,6 +28,9 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if(is_null($this->user) || !$this->user->can('role.view')){
+            abort(403, 'Unauthorized access');
+        }
         $roles = Role::all();
         return view('backend.pages.roles.index',compact('roles'));
     }
@@ -29,6 +42,9 @@ class RolesController extends Controller
      */
     public function create()
     {
+        if(is_null($this->user) || !$this->user->can('role.create')){
+            abort(403, 'Unauthorized access');
+        }
         $permissions = Permission::all();
         $permission_groups = User::getpermissionGroup();
         // dd($permissions_group);
@@ -43,13 +59,16 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if(is_null($this->user) || !$this->user->can('role.create')){
+            abort(403, 'Unauthorized access');
+        }
         $request->validate([
             'name' => 'required|max:100|unique:roles'
         ],[
             'name.required' => 'please write a role name'
         ]);
 
-       $role = Role::create(['name' => $request->name]);
+       $role = Role::create(['name' => $request->name, 'guard_name' => 'admin']);
        $permissions = $request->permissions;
        if(!empty($permissions)){
         $role->syncPermissions($permissions);
@@ -77,7 +96,10 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findById($id);
+        if(is_null($this->user) || !$this->user->can('role.edit')){
+            abort(403, 'Unauthorized access');
+        }
+        $role = Role::findById($id,'admin');
         // dd($role->permissions);
         $all_permission = Permission::all();
         $permission_groups = User::getpermissionGroup();
@@ -93,13 +115,16 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(is_null($this->user) || !$this->user->can('role.update')){
+            abort(403, 'Unauthorized access');
+        }
         $request->validate([
             'name' => 'required|max:100|unique:roles,name,' . $id
         ],[
             'name.required' => 'please write a role name'
         ]);
 
-       $role = Role::findById($id);
+       $role = Role::findById($id,'admin');
        $permissions = $request->permissions;
        if(!empty($permissions)){
         $role->syncPermissions($permissions);
@@ -117,7 +142,10 @@ class RolesController extends Controller
     public function destroy($id)
     {
     //    dd($id);
-        $role = Role::findById($id);
+    if(is_null($this->user) || !$this->user->can('role.delete')){
+        abort(403, 'Unauthorized access');
+    }
+        $role = Role::findById($id,'admin');
         if(!is_null($role)){
             $role->delete();
         }
